requirements.txt
Pillow
torchvision
torch
